requirements.txt
Django==1.6.5
djangorestframework==3.2.3
MySQL-python==1.2.5
Pygments==2.0.2
ipython==2.1.0
South==0.8.4
